Include an ArrayCount file (Just a beginning)
Was included an ArrayCount file to save the ArrayCount variable for the case of lost of energy making Arduino shuts down, this could be a problem to the work of publish with the buffer file, saving the ArrayCount in the SD is a solution.

For now was just included the lines of code that create the file (if it not exists) and to read the file in Arduino's setup.